+++ src/Firebase/Messaging/MessageData.php
@@ -39,6 +39,6 @@ final class MessageData implements \JsonSerializable
 
     private static function isStringable($value): bool
     {
-        return \is_scalar($value) || (\is_object($value) && \method_exists($value, '__toString'));
+        return \is_null($value) || \is_scalar($value) || (\is_object($value) && \method_exists($value, '__toString'));
     }
 }